+++ app/src/main/java/org/finalappproject/findapetsitter/application/FindaPetSitterApplication.java
@@ -11,7 +11,7 @@ import org.finalappproject.findapetsitter.R;
 import org.finalappproject.findapetsitter.model.Address;
 import org.finalappproject.findapetsitter.model.EmergencyContact;
 import org.finalappproject.findapetsitter.model.Pet;
-import org.finalappproject.findapetsitter.model.RequestParseObject;
+import org.finalappproject.findapetsitter.model.Request;
 import org.finalappproject.findapetsitter.model.User;
 
 /**
@@ -28,7 +28,7 @@ public class FindaPetSitterApplication extends Application {
         ParseObject.registerSubclass(Address.class);
         ParseObject.registerSubclass(EmergencyContact.class);
         ParseObject.registerSubclass(Pet.class);
-        ParseObject.registerSubclass(RequestParseObject.class);
+        ParseObject.registerSubclass(Request.class);
 
         // Initialize Parse
         Parse.initialize(new Parse.Configuration.Builder(this)
